requirements.txt
Flask
gunicorn
mysql-connector
Werkzeug
 